requirements.txt
streamlit
tensorflow
numpy
Pillow
streamlit-drawable-canvas